LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/privacy/subframe-cross-origin-frame.https.sub-expected.txt



FAIL subframe loaded cross-origin that itself loads iframes assert_array_equals: Cross origin should not draw expected property 2 to be 255 but got 0 (expected array [0, 0, 255, 255] got object "0,0,0,0")
